Some network table stuff and misc stuff

+++ auto_encoders.cpp
@@ -4,7 +4,7 @@ auto_encoders::auto_encoders(hw_info left1, hw_info right1, hw_info left2, hw_in
     left_encoder = new Encoder(left1.moduleNumber, left1.channel, left2.moduleNumber, left2.channel);
     right_encoder = new Encoder(right1.moduleNumber, right1.channel, right2.moduleNumber, right2.channel);
 }
-double auto_encoders::get_avg_ticks() {
+int auto_encoders::get_avg_ticks() {
     return (left_encoder->Get() + right_encoder->Get()) / 2;
 }
 double auto_encoders::convert_distance(int ticks) {
